添加Real类中的注释

+++ 编译原理/Real.h
@@ -3,10 +3,11 @@
 #include "Tag.h"
 namespace Lexer {
 	// 类 Real 用于处理浮点数
+	//包括float类型和double类型
 	class Real : public Token
 	{
 	public:
-		const float value;
+		const double value;
 
 		Real(float v) : value(v), Token(Tag::REAL) {}
